refactor(user/password-change): change user fields due to new template
